With activity like: Running tool: read_file... Finished tool: read_file
Let's change this so that it's a single entry that gets updated as progress is made. Let's also see details in the summary for which file it was and the directory wrt the project folder. For the expansion, let's see what parts of the file it's trying to read or what parts it does read when it's done.