Ok, let's test things out. Can you plan a task and then execute it. I want to see the lifecycle. ... Your task items were created like: Test task lifecycle pending [ ] [object Object] [ ] [object Object] [ ] [object Object]
First of all, if you are finished, they should no longer be pending. Second, they are ""[object Object]" and properly taking their data. Third, when they are complete they should have a checkmark next to them. Let's review our codebase and come up with an improvement for this process. Then we will execute it.